Time to Commit
As always I've done a shitton of stuff here, and not all of it was
successful. I'm committing now because it's getting messy. I think
I need to move the mouse handling stuff into a meta component.

* Got relationships working to a point
* Added a context menu idea that may not pan out.
* Tried sending mouse events through functions from paper to
children. Sort of worked. Pretty neat. I left the scaffolding there
in case I need it in the future. And frankly I'd like it. The purpose
was to continue to process mouse events even if focus is lost.
I might be able to make it work after my refactor.