Well, it's done.
Paper handles all the mouse input. Things fly away on the screen!
Unfortunately paper is now a shithole of code. The other components
are nice and tidy though!

Also, the object editor dialog is broken. It needs to be moved to paper
I guess. Which will only make it even uglier.